call_forward/requirements.txt
azure-functions
azure-communication-callautomation
